% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{AbsorptionAlphaFisherSimmons}
\alias{AbsorptionAlphaFisherSimmons}
\title{Calculation of absorption in sea water from Fisher and Simmons 1977}
\source{
Fisher and Simmons, 1977
J. Acoust. Soc. Am., Vol. 62, No. 3, September 1977
}
\usage{
AbsorptionAlphaFisherSimmons(f, temperatureC, D)
}
\arguments{
\item{f,}{frequency (kHz)}

\item{temperatureC,}{temperature in degrees C}

\item{D,}{depth in meters}
}
\value{
the absorption
}
\description{
Returns the absorption in sea water from Fisher and Simmons 1977
}
\examples{
AbsorptionAlphaFisherSimmons(20, 0, 1)
}
\author{
Jose Gama
}

