% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{AbsorptionSoundFreshWaterFrancoisGarrison}
\alias{AbsorptionSoundFreshWaterFrancoisGarrison}
\title{Calculation of absorption of sound in fresh water From Francois & Garrison 1982}
\source{
Echoview, 2016
Sonar calculator algorithms
\url{http://support.echoview.com/WebHelp/Reference/Algorithms/Sonar_calculator_algorithms.htm}
}
\usage{
AbsorptionSoundFreshWaterFrancoisGarrison(SonarFreq, temperatureC, D)
}
\arguments{
\item{SonarFreq,}{sonar frequency (kHz)}

\item{temperatureC,}{temperature (degC)}

\item{D,}{depth in meters}
}
\value{
the absorption of sound
}
\description{
Returns the absorption of sound in fresh water From Francois & Garrison 1982
Total absorption = Pure Water Contrib.
}
\examples{
AbsorptionSoundFreshWaterFrancoisGarrison(50, 0, 0)
}
\author{
Jose Gama
}
\references{
Francois & Garrison 1982
Sound absorption based on ocean measurements:
Part I:Pure water and magnesium sulfate contributions
J. Acoust. Soc. Am., Vol. 72, No. 6
}

