% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Analyse_komplett.R
\name{analyze.ddf}
\alias{analyze.ddf}
\title{Analyze complete .ddf-file}
\usage{
analyze.ddf(my.file, win.start = 0.83, win.length = 5, vers = 3,
  y.lim = 512, max.frame, a.1 = 18, a.2 = 0, cut = 50, m.d.cs = 0.2,
  pfad.mult, n.angle = 3, signal.neg = FALSE, do.plot = FALSE,
  do.plot.vec = FALSE, do.plot.hots = FALSE, do.plot.hots.vecs = FALSE,
  plot.hots.mult = FALSE, n.cores = 1, df.t = NULL, frames.pack = 600,
  maxdist = 5, fastclust = FALSE, speeditup = FALSE, floodclust = FALSE,
  floodCpp = TRUE, only.watch = TRUE, schwarm.find = FALSE,
  regel.schwarm = NULL, which.regel.schwarm = NULL,
  save.preprocess = FALSE, zeitstempel, min.length = 10, fisch = "Object",
  max.row = 50, center = FALSE, long = TRUE, file.regel, klass.regel,
  sdcorr = FALSE, FUN = mean, FUN.ow = mean, save.jpg = TRUE,
  save.movie = TRUE, delete.jpg = TRUE, each = 5, wait.jpg = FALSE,
  win = FALSE)
}
\arguments{
\item{my.file}{Filename of .ddf file to be analyzed}

\item{win.start}{Start of the sonar window, possibly extracted before via 
\code{\link{get.version()}}}

\item{win.length}{Length of the sonar window, possibly extracted before via 
\code{\link{get.version}}}

\item{vers}{Version of the .ddf file, possibly extracted before via 
\code{\link{get.version()}}}

\item{y.lim}{Number of pixels to analyze in each beam, counted from the 
camera}

\item{max.frame}{Total number of frames \code{max.frame} of the given video, 
possibly extracted before via \code{\link{get.version()}}}

\item{a.1}{Threshold for centered data}

\item{a.2}{Threshold for uncentered data, default \code{0} means that no thresholding is done}

\item{cut}{Minimal size of cluster in pixels}

\item{m.d.cs}{Maximal distance for which two hotspots can be assigned the 
same tracking number}

\item{pfad.mult}{Base folder for resulting plots and output}

\item{n.angle}{Number of watch hands per quarter, i.e., total number of watch hands is \code{4 x n.angle}}

\item{signal.neg}{Is the signal negativ? Default is \code{FALSE}}

\item{do.plot}{\code{TRUE}: Plots are saved, default is \code{FALSE}}

\item{do.plot.vec}{\code{TRUE}: Plots are saved, default is \code{FALSE}}

\item{do.plot.hots}{\code{TRUE}: Plots are saved, default is \code{FALSE}}

\item{do.plot.hots.vecs}{\code{TRUE}: Plots are saved, default is \code{FALSE}}

\item{plot.hots.mult}{\code{TRUE}: Plots are saved, default is \code{FALSE}}

\item{n.cores}{Number of cores if parallelization is needed, default is \code{1}}

\item{df.t}{Number of degrees of freedom for the splines. Default \code{NULL} 
results in \code{c(100, 25, round(n.frames/2))} where \code{n.frames} is the 
total number of frames}

\item{frames.pack}{Number of frames to be analyzed at once, possibly the 
total number of frames \code{max.frame} of the given video is 
extracted before via \code{\link{get.version()}}}

\item{maxdist}{Maximal distance of two pixels to be assigned to the same cluster, 
not relevant for cluster methods floodclust and floodCpp}

\item{fastclust}{\code{TRUE}: fastclust method is used for clustering of 
pixels and definition of hotspots, default is \code{FALSE}}

\item{speeditup}{\code{TRUE}: fast version of fastclust method is used for clustering of 
pixels and definition of hotspots, default is \code{FALSE}}

\item{floodclust}{\code{TRUE}: floodclust method is used for clustering of 
pixels and definition of hotspots, default is \code{FALSE}}

\item{floodCpp}{\code{TRUE}: floodCpp method is used for clustering of 
pixels and definition of hotspots, default is \code{TRUE} - recommended method}

\item{only.watch}{\code{TRUE}: Use watch hands instead of trapezoid variables - 
strongly recommended, default is \code{TRUE}}

\item{schwarm.find}{\code{TRUE}: A detection of shoal of fish is carried out, 
default is \code{FALSE}}

\item{regel.schwarm}{File name of classification rule for shoal of fish}

\item{which.regel.schwarm}{Type of classification rule for shoal of fish, i.e., lda, qda...}

\item{save.preprocess}{\code{TRUE}: Results of preprocessing are saved, default is \code{FALSE}}

\item{zeitstempel}{time stemp of .ddf file}

\item{min.length}{Minimal length of tracks. Tracks with less hotspots are deleted.}

\item{fisch}{Name of fish for plots}

\item{max.row}{Maximal number of rows in plot of hotspots}

\item{center}{\code{TRUE}: Center variables on hotspot level, default is \code{FALSE}}

\item{long}{\code{TRUE}: Average of all hotspot variables are used, default is \code{TRUE}}

\item{file.regel}{File name of classification rule}

\item{klass.regel}{Type of classification rule for shoal of fish, i.e., lda, qda...}

\item{sdcorr}{\code{TRUE}: Standard deviations and correlations of variables 
on hotspot level are used, default is \code{FALSE}}

\item{FUN}{Function for gathering hotspot information on object level, default is \code{mean}}

\item{FUN.ow}{Function for gathering hotspot information on object level 
when not using the trapezoid variables, default is \code{mean}}

\item{save.jpg}{\code{TRUE}: jpgs for the creating of the video are saved, 
default is \code{TRUE}}

\item{save.movie}{\code{TRUE}: Video is created, 
default is \code{TRUE}}

\item{delete.jpg}{\code{TRUE}: jpgs for the creating of the video are deleted 
after creating the video, default is \code{TRUE}}

\item{each}{Interval of images for the video, for example default \code{5} 
means, that each 5th image is saved in the video}

\item{wait.jpg}{\code{TRUE}: Code pauses until creation of video is finished, 
default is \code{FALSE}}

\item{win}{Specify \code{TRUE} if you are running under windows, then, 
parallelization is not possible}
}
\value{
As output, a table is generated which lists the found objects with 
 predicted class and computed features at \code{pfad.mult}
}
\description{
Analyze complete .ddf-file
}

