% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Analyse_komplett.R
\name{analyze_ddf}
\alias{analyze_ddf}
\title{Analyze complete .ddf file}
\usage{
analyze_ddf(ddf.file, win.start = 0.83, win.length = 5, vers = 3,
  max.frame, y.lim = 512, a = 18, cut = 50, m.d.cs = 0.2, folder.output,
  n.angle = 3, do.plots = FALSE, n.cores = 1, frames.pack = 600,
  df.t = NULL, save.preprocess = FALSE, timestamp, min.length = 10,
  file.classrule, type.classrule, save.movie = TRUE, each = 5,
  win = FALSE)
}
\arguments{
\item{ddf.file}{Filename of .ddf file to be analyzed}

\item{win.start}{Start of the sonar window, possibly extracted before via 
\code{\link{get.version}}}

\item{win.length}{Length of the sonar window, possibly extracted before via 
\code{\link{get.version}}}

\item{vers}{Version of the .ddf file, possibly extracted before via 
\code{\link{get.version}}}

\item{max.frame}{Total number of frames of the given video, 
possibly extracted before via \code{\link{get.version}}}

\item{y.lim}{Number of pixels to analyze in each beam, counted from the 
camera}

\item{a}{Threshold for centered data}

\item{cut}{Minimal size of clusters in number of pixels. Smaller clusters are 
deleted before the tracking and hence not considered as hotspots.}

\item{m.d.cs}{Maximal distance for which two hotspots can be assigned the 
same tracking number}

\item{folder.output}{Folder for resulting plots and output}

\item{n.angle}{Number of watch hands per quarter, i.e., total number of 
watch hands is \code{4 x n.angle}}

\item{do.plots}{\code{TRUE}: Plots of preprocessing are saved, default is \code{FALSE}}

\item{n.cores}{Number of cores if parallelization is needed, default is \code{1}}

\item{frames.pack}{Number of frames to be analyzed at once, possibly the 
total number of frames \code{max.frame} of the given video. If \code{frames.pack} is 
  less than \code{max.frame}, the video is splitted in parts of 
  \code{frames.pack} frames and each part is analyzed separately.}

\item{df.t}{Number of degrees of freedom for the splines. Default \code{NULL} 
results in \code{c(100, 25, round(n.frames/2))} where \code{n.frames} is the 
total number of frames of the analyzed video, i.e. \code{length(frames.pack)}.}

\item{save.preprocess}{\code{TRUE}: Results of preprocessing are saved, 
default is \code{FALSE}}

\item{timestamp}{Time stamp of .ddf file}

\item{min.length}{Minimal length of tracks. Tracks with less hotspots are deleted.}

\item{file.classrule}{File name of classification rule}

\item{type.classrule}{Type of classification rule, i.e., lda, qda...}

\item{save.movie}{\code{TRUE}: Video of raw signal and cleaned signal is created, 
default is \code{TRUE}}

\item{each}{Interval of images for the video, for example default \code{5} 
means, that each \code{5}th image is saved in the video}

\item{win}{Specify \code{TRUE} if you are running under windows, then 
parallelization is not possible}
}
\value{
As output, a table is generated which lists the found objects with 
 predicted class and computed features at \code{folder.output}.
}
\description{
This function analyzes a given .ddf file, i.e., carries out the entire 
 preprocessing with identification of hotspots, tracking, extraction of 
 features and finally classifies the identified objects.
}

