% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{BandLevelFlatSpectrum}
\alias{BandLevelFlatSpectrum}
\title{band level (BL) for flat spectrum}
\usage{
BandLevelFlatSpectrum(SpL, deltaf)
}
\arguments{
\item{SpL}{spectrum level}

\item{deltaf}{band frequency}
}
\value{
band level (BL)
}
\description{
Returns the total intensity of the sound in a band for flat spectrum
}
\examples{
BandLevelFlatSpectrum( 3, 2 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 10.
}

