% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{BandLevelFromCompleteBand}
\alias{BandLevelFromCompleteBand}
\title{band level (BL) from complete band}
\usage{
BandLevelFromCompleteBand(I0, f1, f2)
}
\arguments{
\item{I0}{spectrum level}

\item{f1}{lower frequency}

\item{f2}{upper frequency}
}
\value{
band level (BL)
}
\description{
Returns the band level from integrating the intensity over the complete band
}
\examples{
BandLevelFromCompleteBand( 10000, 40000, 50000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 10.
}

