% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{BasicActiveSonarEquation}
\alias{BasicActiveSonarEquation}
\title{basic active sonar equation}
\usage{
BasicActiveSonarEquation(SL, TS, PL, N, DT)
}
\arguments{
\item{SL}{is the source level of the target}

\item{TS}{target strength}

\item{PL}{propagation loss}

\item{N}{noise}

\item{DT}{detection threshold}
}
\value{
SE signal excess (dB)
}
\description{
Returns the basic active sonar equation SE = (SL + TS - 2 * PL) - N - DT
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 120.
}

