% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{BasicPassiveSonarEquation}
\alias{BasicPassiveSonarEquation}
\title{basic passive sonar equation}
\usage{
BasicPassiveSonarEquation(SL, PL, N)
}
\arguments{
\item{SL}{is the source level of the target}

\item{PL}{propagation loss}

\item{N}{noise}
}
\value{
SE signal excess (dB)
}
\description{
Returns the basic passive sonar equation SE = (SL - PL) - N = DT
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 120.
}

