% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{BasicSonarEquation}
\alias{BasicSonarEquation}
\title{basic sonar equation}
\usage{
BasicSonarEquation(S, N, DT)
}
\arguments{
\item{S}{signal}

\item{N}{noise}

\item{DT}{detection threshold}
}
\value{
SE signal excess (dB)
}
\description{
Returns the basic sonar equation SE = S - N + DT
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 120.
}

