% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{CavitationThresholdEstimateFunctionOfDepth}
\alias{CavitationThresholdEstimateFunctionOfDepth}
\title{Cavitation threshold estimate as a function of depth}
\usage{
CavitationThresholdEstimateFunctionOfDepth(d)
}
\arguments{
\item{d,}{depth (meters)}
}
\value{
Cavitation threshold
}
\description{
Returns the Cavitation threshold estimate as a function of depth
line passing by (5, 2) and (50, 50)
}
\examples{
CavitationThresholdEstimateFunctionOfDepth( 1 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 6.
}

