% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{CavitationThresholdEstimateFunctionOfRadiatedAcousticPowerIntensity}
\alias{CavitationThresholdEstimateFunctionOfRadiatedAcousticPowerIntensity}
\title{Cavitation threshold estimate as a function of radiated acoustic power intensity}
\usage{
CavitationThresholdEstimateFunctionOfRadiatedAcousticPowerIntensity(Ir)
}
\arguments{
\item{Ir,}{radiated acoustic power intensity}
}
\value{
Cavitation threshold
}
\description{
Returns the Cavitation threshold estimate as a function of radiated acoustic power intensity
line passing by (2, 5) and (50, 50)
}
\examples{
CavitationThresholdEstimateFunctionOfRadiatedAcousticPowerIntensity ( 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 6.
}

