% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\docType{data}
\name{CorrectiveTermsPressureFromDepth}
\alias{CorrectiveTermsPressureFromDepth}
\title{Corrective terms to be subtracted for obtaining pressure from depth}
\format{dataframe with 14 rows and 5 columns:
\describe{
 \item{No}{Number}
 \item{Area.of.applicability}{Area of applicability}
 \item{Expression.for.deltaf}{Expression for deltaf}
 \item{Latitude}{Latitude in degrees}
 \item{Accuracy}{Accuracy}
}}
\usage{
CorrectiveTermsPressureFromDepth
}
\description{
Corrective terms to be added for obtaining pressure (MPa) from depth (m)
}
\author{
Jose Gama
}
\references{
C. C. Leroy and F Parthiot, 1998
Depth-pressure relationship in the oceans and seas. J. Acoust. Soc. Am. 103(3) pp 1346-1352
}
\keyword{data}

