% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{CutoffFrequencyShallowWater}
\alias{CutoffFrequencyShallowWater}
\title{Calculation of cutoff frequency in shallow water from Jensen et Al 2011}
\source{
Finn B. Jensen, William A. Kuperman, Michael B. Porter, Henrik Schmidt, 2011
Computational Ocean Acoustics, 2nd Edition. Springer. pp. 29
}
\usage{
CutoffFrequencyShallowWater(Cw, Cb, D)
}
\arguments{
\item{Cw,}{velocity of sound in water}

\item{Cb,}{velocity of sound in  homogeneous bottom}

\item{D,}{depth in meters}
}
\value{
the cutoff frequency (Hz)
}
\description{
Returns the cutoff frequency in shallow water from Jensen et Al 2011
}
\examples{
CutoffFrequencyShallowWater(3000, 2500, 1)
}
\author{
Jose Gama
}

