% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{DepthToPressureLeroyParthiot}
\alias{DepthToPressureLeroyParthiot}
\title{Depth To Pressure from Leroy Parthiot 1998}
\source{
C. C. Leroy and F Parthiot, 1998
Underwater Acoustics Technical Guides - Speed of Sound in Sea Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundseawater/}
}
\usage{
DepthToPressureLeroyParthiot(D, latitude, CorrectiveTerm = NA)
}
\arguments{
\item{D,}{depth in meters}

\item{latitude,}{latitude in degrees}

\item{CorrectiveTerm,}{optional corrective term}
}
\value{
the Pressure
}
\description{
Returns the Depth To Pressure from Leroy Parthiot 1998
}
\examples{
DepthToPressureLeroyParthiot(0, 0)
}
\author{
Jose Gama
}
\references{
C. C. Leroy and F Parthiot, 1998
Depth-pressure relationship in the oceans and seas (1998)
J. Acoust. Soc. Am. 103(3) pp 1346-1352
}

