% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{DetectionIndex}
\alias{DetectionIndex}
\title{Detection index}
\usage{
DetectionIndex(S, N)
}
\arguments{
\item{S}{signal}

\item{N}{noise}
}
\value{
Detection index
}
\description{
Returns the Detection index
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 120.
}

