% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Output_erstellen.R
\name{generate.output}
\alias{generate.output}
\title{Generate output of fish classification}
\usage{
generate.output(do.classification, which.select, pred.class, post.obj,
  sum.prepro.roh, frames.pack, track.eval.out.list, merkmale.out.list,
  max.frame, pfad.mult, zeitstempel, vars.hot.list, is.schwarm.vec,
  only.schwarm = FALSE)
}
\arguments{
\item{do.classification}{\code{TRUE}: Results of classification are available}

\item{which.select}{Vector of indices of objects}

\item{pred.class}{Predicted classes}

\item{post.obj}{Posterior probabilities}

\item{sum.prepro.roh}{Summary of the preprocessing, output of 
\code{\link{summary_func}}}

\item{frames.pack}{Number of frames to be analyzed in one package}

\item{track.eval.out.list}{Tracking result, output of \code{\link{preprocess.data}}}

\item{merkmale.out.list}{Features, output of \code{\link{preprocess.data}}}

\item{max.frame}{Total number of frames \code{max.frame} of the given video, 
possibly extracted before via \code{\link{get.version}}}

\item{pfad.mult}{Folder for resulting plots and output}

\item{zeitstempel}{time stamp of .ddf file}

\item{vars.hot.list}{Variables on hotspot level, output of \code{\link{preprocess.data}}}

\item{is.schwarm.vec}{Vector regarding shoal of fish}

\item{only.schwarm}{\code{TRUE}: Only shoal analysis was carried out, 
default is \code{FALSE}}
}
\value{
Output table with class of fish and all features is saved at 
\code{pfad.mult}
}
\description{
This function generates the output of the entire fish classification
}

