% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{InternationalFormulaForGravity}
\alias{InternationalFormulaForGravity}
\title{International Formula For Gravity}
\source{
Fofonoff and R.C. Millard, 1983
Underwater Acoustics Technical Guides - Speed of Sound in Sea Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundseawater/}
}
\usage{
InternationalFormulaForGravity(latitude, CorrectiveTerm = NA)
}
\arguments{
\item{latitude,}{latitude in degrees}

\item{CorrectiveTerm,}{optional corrective term}
}
\value{
the average gravity
}
\description{
Returns the average gravity at certain latitude
}
\examples{
InternationalFormulaForGravity(0)
}
\author{
Jose Gama
}
\references{
Saunders P.M., Fofonoff N.P., 1976
Conversion of pressure to depth in the ocean. Deep Sea Research 23:109-111.
}

