% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{MaximumRadiatedPowerToAvoidCavitation}
\alias{MaximumRadiatedPowerToAvoidCavitation}
\title{Maximum radiated power to avoid cavitation}
\usage{
MaximumRadiatedPowerToAvoidCavitation(radiatingSurfaceArea, cavitationThreshold)
}
\arguments{
\item{radiatingSurfaceArea,}{Radiating surface area}

\item{cavitationThreshold,}{Cavitation threshold}
}
\value{
Maximum radiated power
}
\description{
Returns the Maximum radiated power to avoid cavitation
}
\examples{
MaximumRadiatedPowerToAvoidCavitation( 50, 0.7 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 5.
}

