% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\docType{data}
\name{MolecularRelaxationAttenuationCoeficient}
\alias{MolecularRelaxationAttenuationCoeficient}
\title{Molecular relaxation attenuation coeficient (alpha)}
\format{dataframe with 3 rows and 11 columns:
\describe{
\item{temperatureC}{numeric, temperature in degrees Celsius}
\item{0.5}{attenuation coeficient for 0.5 kHz}
\item{1}{attenuation coeficient for 1 kHz}
\item{2}{attenuation coeficient for 2 kHz}
\item{5}{attenuation coeficient for 5 kHz}
\item{10}{attenuation coeficient for 10 kHz}
\item{20}{attenuation coeficient for 20 kHz}
\item{50}{attenuation coeficient for 50 kHz}
\item{100}{attenuation coeficient for 100 kHz}
\item{200}{attenuation coeficient for 200 kHz}
\item{500}{attenuation coeficient for 500 kHz}
}}
\usage{
MolecularRelaxationAttenuationCoeficient
}
\description{
Returns the attenuation coeficient of absorption losses
due to molecular relaxation
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 47.
}
\keyword{data}

