% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{MolecularRelaxationAttenuationCoeficientApproximation}
\alias{MolecularRelaxationAttenuationCoeficientApproximation}
\title{Molecular relaxation attenuation coeficient approximation}
\usage{
MolecularRelaxationAttenuationCoeficientApproximation(f)
}
\arguments{
\item{f,}{frequency (Hz)}
}
\value{
alpha Molecular relaxation attenuation coeficient
}
\description{
Returns the Molecular relaxation attenuation coeficient approximation
}
\examples{
MolecularRelaxationAttenuationCoeficientApproximation( 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 47.
}

