% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PeakTS}
\alias{PeakTS}
\title{peak pressure of the incident and reflected pulses}
\usage{
PeakTS(Pr, Pi)
}
\arguments{
\item{Pr}{pressure of the reflected pulse}

\item{Pi}{pressure of the incident pulse}
}
\value{
Target Strength (TS)
}
\description{
Returns the peak pressure of the incident and reflected pulses
}
\examples{
PeakTS( 900, 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 68.
}

