% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PlaneWaveIntensity}
\alias{PlaneWaveIntensity}
\title{Plane wave intensity}
\usage{
PlaneWaveIntensity(p, rho, C)
}
\arguments{
\item{p,}{pressure (Pa or N/m2)}

\item{rho,}{fluid density = 10^3kg/m^3 for sea water}

\item{C,}{velocity of sound wave propagation = 1.5 x 10^3m/s in sea water}
}
\value{
intensity of the wave (power / unit area) (Watt / m^2)
}
\description{
Returns the intensity
}
\examples{
PlaneWaveIntensity( 1e3, 1.5e3,  1)
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 3.
}

