% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PLcylindricalSpreadingLaw}
\alias{PLcylindricalSpreadingLaw}
\title{PL to range r cylindrical spreading law in logarithmic form}
\usage{
PLcylindricalSpreadingLaw(r)
}
\arguments{
\item{r}{radius (meters)}
}
\value{
Propagation loss (PL) (dB)
}
\description{
Returns the PL to range r cylindrical spreading law in logarithmic form
}
\examples{
PLcylindricalSpreadingLaw( 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 45.
}

