% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_methods.R
\name{plot.sonar}
\alias{plot.sonar}
\title{Method for plotting 'sonar' objects}
\usage{
\method{plot}{sonar}(x, n = 200)
}
\arguments{
\item{'sonar'}{object}
}
\value{
plot
}
\description{
Returns a plot of the path using Longitude and Latitude
}
