% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_methods.R
\name{plot.sonaR_frames}
\alias{plot.sonaR_frames}
\title{Function to plot a sonaR_frames object}
\usage{
plot.sonaR_frames(frame_list)
}
\arguments{
\item{frame_list}{sonaR_frames. A sonaR_frames object, which is a list of matrixes containing sonar data}
}
\value{
plot
}
\description{
Returns image plot with range on y-axis
}
