% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Preprocessing.R
\name{plot_prepro}
\alias{plot_prepro}
\title{Plot results of the preprocessing}
\usage{
plot_prepro(hot = TRUE, fisch, pfad, sum.prepro, min.length, cart.coord,
  hots.y01.list, dims.list, track.ddf.list, merkmale.out.list, res = 200,
  cex = 1, filename = NULL, max.row = 50, n.cores = 1, a.2)
}
\arguments{
\item{hot}{\code{TRUE} (default): Hotspots are plotted, else watch hands}

\item{fisch}{Species of fish}

\item{pfad}{Folder where the plots will be saved}

\item{sum.prepro}{Results of \code{\link{summary_func}}}

\item{min.length}{Minimal length of tracks}

\item{cart.coord}{Cartesian coordinates of all pixel of the sonar window}

\item{hots.y01.list}{List with hotspots}

\item{dims.list}{List with dimensions of hotspots}

\item{track.ddf.list}{List with tracks}

\item{merkmale.out.list}{List with features}

\item{res}{Resolution of plots (deafault is \code{200})}

\item{cex}{Cex for plotting (default is \code{1})}

\item{filename}{Filename}

\item{max.row}{Maximal number of rows for plots}

\item{n.cores}{Number of cores if parallelization is needed, default is \code{1}}

\item{a.2}{Threshold for uncentered data, default \code{0} means that no thresholding is done}
}
\description{
Plot results of the preprocessing
}

