% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PLsphericalSpreadingLaw}
\alias{PLsphericalSpreadingLaw}
\title{PL to range r spherical spreading law in logarithmic form}
\usage{
PLsphericalSpreadingLaw(r)
}
\arguments{
\item{r}{radius (meters)}
}
\value{
Propagation loss (PL) (dB)
}
\description{
Returns the PL to range r spherical spreading law in logarithmic form
}
\examples{
PLsphericalSpreadingLaw( 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 44.
}

