% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PowerCylindricalSpreadingLaw}
\alias{PowerCylindricalSpreadingLaw}
\title{Power cylindrical spreading law}
\usage{
PowerCylindricalSpreadingLaw(r, h, Ir)
}
\arguments{
\item{r}{radius (meters)}

\item{h}{distance between 2 planes (meters)}

\item{Ir}{intensity at radius r}
}
\value{
total power (Watts)
}
\description{
Returns the Power cylindrical spreading law
}
\examples{
PowerCylindricalSpreadingLaw( 1000, 100, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 45.
}

