% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PowerSphericalSpreadingLaw}
\alias{PowerSphericalSpreadingLaw}
\title{Power spherical spreading law}
\usage{
PowerSphericalSpreadingLaw(r, Ir)
}
\arguments{
\item{r}{radius (meters)}

\item{Ir}{intensity at radius r}
}
\value{
total power (Watts)
}
\description{
Returns the Power spherical spreading law
}
\examples{
PowerSphericalSpreadingLaw( 1000, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 44.
}

