% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Daten_Klassifikation.R
\name{prepare.and.count}
\alias{prepare.and.count}
\title{Prepare data for counting and count}
\usage{
prepare.and.count(track.ddf.list, track.eval.out.list, merkmale.out.list,
  t.grid.list, n.angle, min.length, center, long, only.watch,
  file.regel = "Regel_2.RData", klass.regel, sdcorr, FUN, FUN.ow,
  lda.obj = NULL, qda.obj = NULL, svm.obj = NULL)
}
\arguments{
\item{track.ddf.list}{Tracking, output of \code{\link{preprocess.data}}}

\item{track.eval.out.list}{Tracking evaluation, output of \code{\link{preprocess.data}}}

\item{merkmale.out.list}{Features, output of \code{\link{preprocess.data}}}

\item{t.grid.list}{Grids of time points, output of \code{\link{preprocess.data}}}

\item{n.angle}{Number of watch hands per quarter, i.e., total number of watch hands is \code{4 x n.angle}}

\item{min.length}{Minimal length of tracks. Tracks with less hotspots are deleted.}

\item{center}{\code{TRUE}: Center variables on hotspot level, default is \code{FALSE}}

\item{long}{\code{TRUE}: Average of all hotspot variables are used, default is \code{TRUE}}

\item{only.watch}{\code{TRUE}: Use watch hands instead of trapezoid variables - 
strongly recommended, default is \code{TRUE}}

\item{file.regel}{File name of classification rule for fish species}

\item{klass.regel}{Type of classification rule for fish species, i.e., lda, qda...}

\item{sdcorr}{\code{TRUE}: Standard deviations and correlations of variables 
on hotspot level are used, default is \code{FALSE}}

\item{FUN}{Function for gathering hotspot information on object level, default is \code{mean}}

\item{FUN.ow}{Function for gathering hotspot information on object level 
when not using the trapezoid variables, default is \code{mean}}

\item{lda.obj}{Object of learned classification rule if lda is used, is 
loaded from \code{file.regel}}

\item{qda.obj}{Object of learned classification rule if qda is used, is 
loaded from \code{file.regel}}

\item{svm.obj}{Object of learned classification rule if svm is used, is 
loaded from \code{file.regel}}
}
\value{
Result of classification
}
\description{
This function prepares the output of the previous steps (features...) for
the classification, classifies the objects and counts number of objects
per class
}

