% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PressureBlackSeaSimplifiedLeroy}
\alias{PressureBlackSeaSimplifiedLeroy}
\title{Calculation of pressure in the Black Sea from Leroy 1969}
\source{
Leroy C. C. 1969
Development of simple equations for accurate and more realistic
calculations of the speed of sound in sea water
J. Acoust. Soc. Am. 46, 216-226.
}
\usage{
PressureBlackSeaSimplifiedLeroy(Z, lat)
}
\arguments{
\item{Z,}{depth in meters}

\item{lat,}{latitude n degrees}
}
\value{
the pressure
}
\description{
Returns the pressure in the Black Sea from Leroy 1969
}
\examples{
PressureBlackSeaSimplifiedLeroy(0, 0)
}
\author{
Jose Gama
}

