% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PressureModifiedSimplifiedLeroy}
\alias{PressureModifiedSimplifiedLeroy}
\title{Calculation of pressure in water (Leroy modified) from Lovett 1978}
\source{
Lovett, J.R. 1978
Merged seawater sound-speed equations
J. Acoust. Soc. Am., 63, 1713-18.
}
\usage{
PressureModifiedSimplifiedLeroy(Z, lat)
}
\arguments{
\item{Z,}{depth in meters}

\item{lat,}{latitude n degrees}
}
\value{
the pressure
}
\description{
Returns the pressure in water (Leroy simplified modified)
from Lovett 1978
}
\examples{
PressureModifiedSimplifiedLeroy(0, 0)
}
\author{
Jose Gama
}

