% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PressureToDepthLeroyParthiot}
\alias{PressureToDepthLeroyParthiot}
\title{Pressure To Depth from Leroy Parthiot 1998}
\source{
C. C. Leroy and F Parthiot, 1998
Underwater Acoustics Technical Guides - Speed of Sound in Sea Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundseawater/}
}
\usage{
PressureToDepthLeroyParthiot(P, latitude, CorrectiveTerm = NA)
}
\arguments{
\item{P,}{pressure in MPa (relative to atmospheric pressure)}

\item{latitude,}{latitude in degrees}

\item{CorrectiveTerm,}{optional corrective term}
}
\value{
the depth
}
\description{
Returns the Pressure To Depth from Leroy Parthiot 1998
}
\examples{
PressureToDepthLeroyParthiot(0.1, 0)
}
\author{
Jose Gama
}
\references{
C. C. Leroy and F Parthiot, 1998
Depth-pressure relationship in the oceans and seas (1998)
J. Acoust. Soc. Am. 103(3) pp 1346-1352
}

