% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_methods.R
\name{print.sonar}
\alias{print.sonar}
\title{Method to print 'sonar' objects}
\usage{
\method{print}{sonar}(x)
}
\arguments{
\item{'sonar'}{object}
}
\description{
Print 'sonar' object and report some statistics.
}
