% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{ProjectorSensitivityPower}
\alias{ProjectorSensitivityPower}
\title{Projector Sensitivity Power}
\usage{
ProjectorSensitivityPower(I1, Ir, P)
}
\arguments{
\item{I1}{intensity of source at standard range}

\item{Ir}{reference intensity}

\item{P}{power (Watt)}
}
\value{
response Sv (dB/V)
}
\description{
Returns the Projector Sensitivity Power
}
\examples{
ProjectorSensitivityPower( 10000, 15000, 0.7 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 9.
}

