% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{ProjectorSensitivityVoltage}
\alias{ProjectorSensitivityVoltage}
\title{Projector Sensitivity Voltage}
\usage{
ProjectorSensitivityVoltage(I1, Ir, v)
}
\arguments{
\item{I1}{intensity of source at standard range}

\item{Ir}{reference intensity}

\item{v}{Voltage}
}
\value{
response Sv (dB/V)
}
\description{
Returns the Projector Sensitivity Voltage
}
\examples{
ProjectorSensitivityVoltage( 10000, 15000, 0.7 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 9.
}

