% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{PropagationLoss}
\alias{PropagationLoss}
\title{Propagation loss (PL)}
\usage{
PropagationLoss(I0, Ir)
}
\arguments{
\item{I0}{intensity of the source to a point one metre from its acoustic centre}

\item{Ir}{is the intensity at the receiver}
}
\value{
Propagation loss (PL) (dB)
}
\description{
Returns the Propagation loss (PL)
}
\examples{
PropagationLoss( 1000, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 43.
}

