% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{RangeResolutionCHIRP}
\alias{RangeResolutionCHIRP}
\title{Sonar Range Resolution CHIRP}
\usage{
RangeResolutionCHIRP(SonarBandwidth, Cw)
}
\arguments{
\item{SonarBandwidth,}{Sonar Bandwidth}

\item{Cw,}{Velocity of sound}
}
\value{
the Sonar Range Resolution
}
\description{
Returns the Sonar Range Resolution CHIRP
}
\examples{
RangeResolutionCHIRP(1, 343)
}
\author{
Jose Gama
}

