% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{RangeResolutionMonotonic}
\alias{RangeResolutionMonotonic}
\title{Sonar Range Resolution for monotonic acoustic systems}
\usage{
RangeResolutionMonotonic(SonarPulseDuration, Cw)
}
\arguments{
\item{SonarPulseDuration,}{Sonar Pulse Duration}

\item{Cw,}{Velocity of sound}
}
\value{
the Sonar Range Resolution
}
\description{
Returns the Sonar Range Resolution for monotonic acoustic systems
}
\examples{
RangeResolutionMonotonic(1, 343)
}
\author{
Jose Gama
}

