% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_sl2_cpp}
\alias{read_sl2_cpp}
\title{A "No Frills" Faster Version in C++}
\usage{
read_sl2_cpp(path)
}
\value{
data frame (tibble)
}
\description{
No error checking (not even for file existence). No fancy names for channel/etc.
No unnesting of validity fields. BUT, it's wicked fast.
}
\examples{
read_sl2_cpp(system.file("exdat", "example.sl2", package="arabia")))
}
