% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_sl3_cpp}
\alias{read_sl3_cpp}
\title{Function to read .sl3 files}
\usage{
read_sl3_cpp(path)
}
\value{
data frame (tibble)
}
\description{
Function to read .sl3 files
}
\examples{
#read_sl3_cpp(system.file("exdat", "example.sl2", package="arabia")))
}
