% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sonar_frame.R
\name{read_sonar_frame}
\alias{read_sonar_frame}
\title{Function to read raw sonar data}
\usage{
read_sonar_frame(path, df, frametype, georeference = FALSE)
}
\arguments{
\item{path}{String. Path to '.sl3' or '.sl2' binary file}

\item{df}{data.frame. Output from read_sonar function}

\item{frametype}{String. Type of frame to be extracted, one of "Primary", "Secondary", "DSI (Downscan)", "Left (Sidescan)", "Right (Sidescan)", "Composite (Sidescan)"}

\item{georeference}{Boolean. Attempts to georeference sonar data using coordinates in header. Only possible for Sidescan data.}
}
\value{
RasterLayer
}
\description{
Read raw data from sonar recordings using a data.frame containing header data.
Returns raster with rawdata.
}
