% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sonar_header.R
\name{read_sonar_header}
\alias{read_sonar_header}
\title{Function to read data stored in recorded sonar files.}
\usage{
read_sonar_header(path, display_progress = TRUE)
}
\arguments{
\item{path}{String. Path to '.sl3' or '.sl2' binary file}

\item{display_progress}{Boolean. Display progress bar?}
}
\value{
data.frame
}
\description{
Read data from sonar. Only '.sl2' and '.sl3' formats by Lowrance are currently supported.
Data are stored as a header followed by a frame containing the raw sonar data.
The function returns a data.frame where each row represents a header.
}
