% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SLdirectionalProjector}
\alias{SLdirectionalProjector}
\title{SL of a directional projector}
\usage{
SLdirectionalProjector(P, DIt)
}
\arguments{
\item{P,}{power output (watts)}

\item{DIt,}{transmit directivity index (dB)}
}
\value{
SL of a directional projector
}
\description{
Returns the SL of a directional projector
}
\examples{
SLdirectionalProjector( 700, 0.7 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 4.
}

