% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SLomnidirectionalProjector}
\alias{SLomnidirectionalProjector}
\title{SL of an omnidirectional projector}
\usage{
SLomnidirectionalProjector(P)
}
\arguments{
\item{P,}{omnidirectional power output (watts)}
}
\value{
source level (SL)
}
\description{
Returns the SL of an omnidirectional projector
}
\examples{
SLomnidirectionalProjector( 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 4.
}

