% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONAR_func.R
\name{SONAR.deconvolute}
\alias{SONAR.deconvolute}
\title{SONAR.deconvolute}
\usage{
SONAR.deconvolute(
  fname,
  path,
  h,
  verbose = TRUE,
  desktop = FALSE,
  splash = FALSE,
  display = FALSE,
  wait = FALSE,
  single_thread = FALSE
)
}
\arguments{
\item{fname}{the path to SONAR's MATLAB core codes}

\item{path}{the path to preprocessed data}

\item{h}{bandwidth}

\item{verbose}{show the command in MATLAB}

\item{desktop}{Determines whether the matlab run process is displayed}

\item{splash}{Determines whether the matlab run process is displayed}

\item{display}{Determines whether the matlab run process is displayed}

\item{wait}{wait the SONAR's deconvolution complete}

\item{single_thread}{Determine whether to use only single threads}
}
\value{
the state of deconvolution
}
\description{
SONAR.deconvolute
}
