% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONAR_func.R
\name{SONAR.deliver}
\alias{SONAR.deliver}
\title{SONAR.deliver}
\usage{
SONAR.deliver(processed_data, path)
}
\arguments{
\item{processed_data}{the data from preprocessed steps}

\item{path}{the path of delivering the preprocessed data to SONAR}
}
\value{
this part is to save the files, return the complete state
}
\description{
SONAR.deliver
}
