% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SONAR_func.R
\name{SONAR.preprocess}
\alias{SONAR.preprocess}
\title{SONAR.preprocess}
\usage{
SONAR.preprocess(
  sc_count,
  sc_cell_type,
  sc_nUMI,
  sp_coords,
  sp_count,
  sp_nUMI,
  cores = 8,
  type_min_cell = 0,
  spot_min_UMI = 100,
  preclus_resolution = 0.7
)
}
\arguments{
\item{sc_count}{scRNA-seq expression matrix of reference(Details and formats are in vignettes)}

\item{sc_cell_type}{cell type annotation for reference(Details and formats are in vignettes)}

\item{sc_nUMI}{cell library size(Details and formats are in vignettes)}

\item{sp_coords}{spot expression coordinate(Details and formats are in vignettes)}

\item{sp_count}{spot expression matrix(Details and formats are in vignettes)}

\item{sp_nUMI}{spot library size(Details and formats are in vignettes)}

\item{cores}{the number of CPU threads that you want to use}

\item{type_min_cell}{Specifies a threshold to filter out cell types with a smaller number of cells}

\item{spot_min_UMI}{Specifies a threshold to filter out spots with fewer UMIs}

\item{preclus_resolution}{Specifies the resolution of the preclustering}
}
\value{
preprocessed data
}
\description{
SONAR.preprocess
}
