% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SONAR.prop}
\alias{SONAR.prop}
\title{SONAR.prop}
\usage{
SONAR.prop(
  scale = F,
  proportion,
  spatial_location,
  select_type,
  colors = c("grey", "red"),
  NumCols
)
}
\arguments{
\item{scale}{if TRUE, the proportion will be scaled by Min-Max Scaling; else use the original predicted proportion}

\item{proportion}{deconvolution results matrix from SONAR(Details and formats are in vignettes)}

\item{spatial_location}{spatial coordinates matrix(Details and formats are in vignettes)}

\item{select_type}{the types you want to visualize}

\item{colors}{color vector that you want to use}

\item{NumCols}{decide the col-numbers you want to arrange}
}
\value{
ggplot object
}
\description{
SONAR.prop
}
