% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SONAR.visualize.pie}
\alias{SONAR.visualize.pie}
\title{SONAR.visualize.pie}
\usage{
SONAR.visualize.pie(proportion, spatial_location, colors = colors)
}
\arguments{
\item{proportion}{deconvolution results matrix from SONAR(Details and formats are in vignettes)}

\item{spatial_location}{spatial coordinates matrix(Details and formats are in vignettes)}

\item{colors}{color vector that you want to use}
}
\value{
ggplot object
}
\description{
SONAR.visualize.pie
}
