% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_funcs.R
\name{sonar_depth_intensity}
\alias{sonar_depth_intensity}
\title{Function to extract raw sonar data at waterdepth}
\usage{
sonar_depth_intensity(sonar, channel, window_size = 0)
}
\arguments{
\item{channel}{Target channel for which intensity at depth should be extracted}

\item{window_size}{Default = 0.}

\item{'sonar'}{object}
}
\value{
'sonar' object with column 'IntensityAtDepth' added
}
\description{
Extracts and returns a new column with the raw sonar intensity at the reported waterdepth
Optionally, a window size can be given to smooth the result.
}
