% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_frames.R
\name{sonar_frames}
\alias{sonar_frames}
\title{Function to extract raw sonar data}
\usage{
sonar_frames(path, sonar_header, frametype, return_df = FALSE)
}
\arguments{
\item{path}{String. Path to '.sl3' or '.sl2' binary file}

\item{sonar_header}{data.frame. Output data frame from sonar_header function containing header data}

\item{frametype}{String. Type of frame to be extracted, one of "Primary", "Secondary", "Downscan", "LeftSidescan", "RightSidescan", "Sidescan"}

\item{return_df}{Boolean. Return list of header data.frames with a list column containing each frame}
}
\value{
sonaR_frames or sonaR_header if return_df = TRUE
}
\description{
Extract raw data from sonar recordings using a data.frame containing header data read from the same log same.
Returns raster with rawdata.
}
