% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_funcs.R
\name{sonar_image}
\alias{sonar_image}
\title{Function to extract raw sonar data as images/matrix}
\usage{
sonar_image(sonar, channel, normalize_sidescan = FALSE)
}
\arguments{
\item{channel}{Target channel which should be converted to matrix}

\item{normalize_sidescan}{Boolean. Normalize sidescan data using the mean intensity for each angle.}

\item{'sonar'}{object}
}
\value{
list of matrices with range attribute
}
\description{
Extracts raw sonar data from 'sonar' object and returns list of matrices that can be plotted.
New matrix is created for each time the range (MinRange and MaxRange attached to matrices as attribute) of the recorded sonar data changes.
}
