% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sonar.R
\name{sonar_read}
\alias{sonar_read}
\title{Read data stored in sonar files.}
\usage{
sonar_read(path, display_progress = TRUE, read_frames = TRUE)
}
\arguments{
\item{path}{String. Path to '.sl3' or '.sl2' binary file}

\item{display_progress}{Boolean. Display progress bar?}

\item{read_frames}{Boolean. Read metadata and frames.}
}
\value{
Object of class sonar.
}
\description{
Function to read recorded data from sonar files.
Only '.sl2' and '.sl3' formats by Lowrance are currently supported.
Data are stored as a header containing metadata for each recording (e.g. coordinates, water tempereature, speed etc.) followed by a frame containing the raw sonar 'ping' data.
All data is returned in one object of type 'sonar' which in essence is a data.frame, where each row represents a recording with frame data stored in a list-column.
}
