% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_funcs.R
\name{sonar_show_image}
\alias{sonar_show_image}
\title{Function to plot sonar data}
\usage{
sonar_show_image(mat_list)
}
\arguments{
\item{list}{List of matrices returned from sonar_image function}
}
\value{
Plot of all images
}
\description{
Plots output from sonar_image function. One plot for each matrix is created.
}
