% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_sidescan_geo.R
\name{sonar_sidescan_geo}
\alias{sonar_sidescan_geo}
\title{Function to extract raw sidescan data and perform georeferencing using points in header}
\usage{
sonar_sidescan_geo(path, sonar_header)
}
\arguments{
\item{path}{String. Path to '.sl3' or '.sl2' binary file}

\item{sonar_header}{data.frame. Output data frame from sonar_header function containing header data}
}
\value{
list containing RasterLayer objects
}
\description{
Extract raw data from sonar recordings using a data.frame containing header data read from the same log same.
Returns raster with rawdata.
}
