% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SonarEquation}
\alias{SonarEquation}
\title{sonar equation}
\usage{
SonarEquation(SL, PL, TS)
}
\arguments{
\item{SL}{source level}

\item{PL}{propagation loss}

\item{TS}{target strength}
}
\value{
EL echo level
}
\description{
Returns the sonar equation EL = SL - 2PL + TS
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 68.
}

