% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SourceLevel}
\alias{SourceLevel}
\title{source level (SL)}
\usage{
SourceLevel(I1, Ir)
}
\arguments{
\item{I1,}{intensity of source at standard range}

\item{Ir,}{reference intensity}
}
\value{
source level (SL)
}
\description{
Returns the source level (SL)
}
\examples{
SourceLevel( 1000, 1100)
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 4.
}

