% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SourceLevelToAvoidCavitation}
\alias{SourceLevelToAvoidCavitation}
\title{source level to avoid cavitation}
\usage{
SourceLevelToAvoidCavitation(f, DIt)
}
\arguments{
\item{f,}{frequency (Hz)}

\item{DIt,}{transmit directivity index (dB)}
}
\value{
source level SL (dB)
}
\description{
Returns the source level to avoid cavitation
}
\examples{
SourceLevelToAvoidCavitation( 20000, 0.7 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 5.
}

