% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\docType{data}
\name{SpeedAlgorithmParameterRanges}
\alias{SpeedAlgorithmParameterRanges}
\title{Data on Speed of Sound Algorithm Parameter Ranges}
\format{dataframe with 10 rows and 10 columns:
\describe{
 \item{Reference}{Reference}
 \item{TemperatureRangeMin}{Temperature Range (C) Min}
 \item{TemperatureRangeMax}{Temperature Range (C) Min}
 \item{SalinityRangeMin}{Salinity Range (ppt) min}
 \item{SalinityRangeMax}{Salinity Range (ppt) max}
 \item{PressureOrDepthRangeMin}{Pressure or Depth Range min}
 \item{PressureOrDepthRangeMax}{Pressure or Depth Range max}
 \item{PressureOrDepthRangeUnits}{Pressure or Depth Range units}
 \item{StandardError}{Standard Error}
 \item{NumberOfTerms}{Number of Terms}
}}
\usage{
SpeedAlgorithmParameterRanges
}
\description{
Data on Speed of Sound Algorithm Parameter Ranges
}
\author{
Jose Gama
}
\references{
Paul C. Etter, 2013
Underwater Acoustic Modeling and Simulation, Fourth Edition
pp. 28. CRC Press
}
\keyword{data}

