% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundHumidAir}
\alias{SpeedOfSoundHumidAir}
\title{Speed of sound in Humid air at sea level air density and one atmosphere pressure}
\usage{
SpeedOfSoundHumidAir(temperatureC, Hr)
}
\arguments{
\item{temperatureC}{numeric, temperature in degrees Celsius in the range -30 C to 43 C}

\item{Hr}{numeric, relative humidity, accurate to within 0.1\% for temperatures}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound in Humid air at sea level air density
and one atmosphere pressure
}
\examples{
#Speed of sound (Humid air, at sea level air density, one atmosphere pressure)
#temperature 20 C, 90\\\% relative humidity
SpeedOfSoundHumidAir(20, 0.9)
}
\author{
Jose Gama
}
\references{
Kleeman L. & Kuc R., 2008
Springer Handbook of Robotics, Bruno Siciliano, Oussama Khatib (Eds.)
Springer-Verlag Berlin Heidelberg, pp.496 eq (21.7).
}

