% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundKinslerEtal}
\alias{SpeedOfSoundKinslerEtal}
\title{Speed of sound (m/s) from Kinsler et al}
\usage{
SpeedOfSoundKinslerEtal(PressureBar, temperatureC)
}
\arguments{
\item{PressureBar,}{Pressure in bars (1 bar = 100 kPa)}

\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound (m/s) from Kinsler et al
accurate to within 0.05\% for 0<=T<=100 C and 0<=P<=200 bar
}
\examples{
SpeedOfSoundKinslerEtal(1, 20)
}
\author{
Jose Gama
}
\references{
L. Kinsler, A. Frey, A. Coppens, J. Sanders, 1982
Fundamentals of Acoustics, Third Edition
New York: John Wiley & Sons. pp. 121 (5.6.8)
}

