% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundPureWaterBilaniukWong112}
\alias{SpeedOfSoundPureWaterBilaniukWong112}
\title{speed of sound (m/s) from Bilaniuk and Wong 112 point equation}
\source{
National Physical Laboratory, 2015
Underwater Acoustics Technical Guides - Speed of Sound in Pure Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundpurewater/content.html#LUBBERS}
}
\usage{
SpeedOfSoundPureWaterBilaniukWong112(temperatureC)
}
\arguments{
\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns returns the speed of sound (m/s)
Range of validity: 0-100 OC at atmospheric pressure
}
\examples{
SpeedOfSoundPureWaterBilaniukWong112(20)
}
\author{
Jose Gama
}
\references{
Bilaniuk and Wong 1993
Speed of sound in pure water as a function of temperature, J. Acoust. Soc. Am. 93(3) pp 1609-1612
Bilaniuk and Wong 1996
Erratum: Speed of sound in pure water as a function of temperature [J. Acoust. Soc. Am. 93, 1609-1612 (1993)], J. Acoust. Soc. Am. 99(5), p 3257.
}

