% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundPureWaterLubbersandGraaffSEa}
\alias{SpeedOfSoundPureWaterLubbersandGraaffSEa}
\title{speed of sound (m/s) from Lubbers and Graaff's simplified equations a and b}
\source{
National Physical Laboratory, 2015
Underwater Acoustics Technical Guides - Speed of Sound in Pure Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundpurewater/content.html#LUBBERS}
}
\usage{
SpeedOfSoundPureWaterLubbersandGraaffSEa(temperatureC)
}
\arguments{
\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns returns the speed of sound (m/s)
temperature interval 15-35 C at atmospheric pressure, maximum error 0.18 m/s
Lubbers and Graaff's simplified equation a)
}
\examples{
SpeedOfSoundPureWaterLubbersandGraaffSEa(20)
}
\author{
Jose Gama
}
\references{
J. Lubbers and R. Graaff, 1998
A simple and accurate formula for the sound velocity in water,
Ultrasound Med. Biol. Vol 24, No 7, pp 1065-1068.
}

