% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundPureWaterMarczak}
\alias{SpeedOfSoundPureWaterMarczak}
\title{speed of sound (m/s) from Marczak}
\source{
National Physical Laboratory, 2015
Underwater Acoustics Technical Guides - Speed of Sound in Pure Water
\url{http://resource.npl.co.uk/acoustics/techguides/soundpurewater/content.html#LUBBERS}
}
\usage{
SpeedOfSoundPureWaterMarczak(temperatureC)
}
\arguments{
\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns returns the speed of sound (m/s)
Range of validity: 0-95C at atmospheric pressure
}
\examples{
SpeedOfSoundPureWaterMarczak(20)
}
\author{
Jose Gama
}
\references{
Marczak, 1997
Water as a standard in the measurements of speed of sound in liquids
J. Acoust. Soc. Am. 102(5) pp 2776-2779.
}

