% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterLeroy68}
\alias{SpeedOfSoundSeaWaterLeroy68}
\title{speed of sound (m/s) in sea water from Leroy 1968}
\source{
Lurton, X, 2002
An Introduction to Underwater Acoustics, 1st ed. London, Praxis Publishing LTD, p37.
}
\usage{
SpeedOfSoundSeaWaterLeroy68(D, latitude)
}
\arguments{
\item{D,}{depth in meters}

\item{latitude,}{latitude in degrees}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound (m/s)
}
\examples{
SpeedOfSoundSeaWaterLeroy68(25, 0)
}
\author{
Jose Gama
}
\references{
Lurton, X, 2002
An Introduction to Underwater Acoustics, 1st ed. London, Praxis Publishing LTD, p37.
}

