% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterLeroy69}
\alias{SpeedOfSoundSeaWaterLeroy69}
\title{speed of sound (m/s) in sea water from Leroy 1969}
\source{
Leroy C.C. 1969
Underwater Acoustics Technical Guides - Speed of Sound in Sea Water
\url{http://support.echoview.com/WebHelp/Reference/Algorithms/Sonar_calculator_algorithms.htm}
}
\usage{
SpeedOfSoundSeaWaterLeroy69(D, S, temperatureC)
}
\arguments{
\item{D,}{depth in meters}

\item{S,}{salinity in parts per thousand}

\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound (m/s)
Range of validity: -2:23 C, Error +-0.1
}
\examples{
SpeedOfSoundSeaWaterLeroy69(0, 35, 25)
}
\author{
Jose Gama
}
\references{
Leroy C.C. 1969
Development of simple equations for accurate and more realistic calculation
of the speed of sound in sea water. J. acoust. Soc. Am., 46, 216-26.
}

