% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterLovett1}
\alias{SpeedOfSoundSeaWaterLovett1}
\title{Calculation of speed of sound in sea water from Lovett 1978}
\source{
Lovett, J.R. 1978
Merged seawater sound-speed equations
J. Acoust. Soc. Am., 63, 1713-18.
}
\usage{
SpeedOfSoundSeaWaterLovett1(temperatureC, S, P)
}
\arguments{
\item{temperatureC,}{temperature in degrees C T48}

\item{S,}{salinity in \%}

\item{P,}{pressure in decibars (0 at surface)}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound in sea water from Lovett 1978
Check value: at T=2C, S=34.7; P=6000 dbar; C = 1559.462 m/s.
}
\examples{
SpeedOfSoundSeaWaterLovett1(2, 34.7, 6000)
}
\author{
Jose Gama
}

