% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterLovett2}
\alias{SpeedOfSoundSeaWaterLovett2}
\title{Calculation of speed of sound in sea water from Lovett 1978b}
\source{
Lovett, J.R. 1978
Merged seawater sound-speed equations
J. Acoust. Soc. Am., 63, 1713-18.
}
\usage{
SpeedOfSoundSeaWaterLovett2(temperatureC, S, P)
}
\arguments{
\item{temperatureC,}{temperature in degrees C T48}

\item{S,}{salinity in \%}

\item{P,}{pressure in decibars (0 at surface)}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound in sea water from Lovett 1978b
Check value: at T=2C, S=34.7; P=6000 dbar; C = 1559.393 m/s.
}
\examples{
SpeedOfSoundSeaWaterLovett2(2, 34.7, 6000)
}
\author{
Jose Gama
}

