% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterMackenzie}
\alias{SpeedOfSoundSeaWaterMackenzie}
\title{speed of sound (m/s) in sea water from Mackenzie 1981}
\source{
Mackenzie K.V., 1981
Underwater Acoustics Technical Guides - Speed of Sound in Sea Water
\url{http://support.echoview.com/WebHelp/Reference/Algorithms/Sonar_calculator_algorithms.htm}
}
\usage{
SpeedOfSoundSeaWaterMackenzie(D, S, temperatureC)
}
\arguments{
\item{D,}{depth in meters}

\item{S,}{salinity in parts per thousand}

\item{temperatureC,}{temperatureC in Celsius}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound (m/s)
}
\examples{
SpeedOfSoundSeaWaterMackenzie(0, 35, 25)
}
\author{
Jose Gama
}
\references{
Mackenzie K.V., 1981
Nine-term equation for sound speed in the ocean.
J. acoust. Soc. Am., 70, 807-12.
}

