% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{SpeedOfSoundSeaWaterSkone}
\alias{SpeedOfSoundSeaWaterSkone}
\title{speed of sound (m/s) in sea water from Skone et al 2002}
\source{
de Jong, C.D., Lachapelle, G., Skone, S. and Elema, I. A., 2002
 Hydrography. Delft University Press (The Netherlands). pp.194
}
\usage{
SpeedOfSoundSeaWaterSkone(temperatureC, D, S)
}
\arguments{
\item{temperatureC,}{temperature in degrees Celsius}

\item{D,}{depth in meters}

\item{S,}{salinity in parts per thousand}
}
\value{
the speed of sound (m/s)
}
\description{
Returns the speed of sound (m/s)
 modelled using empirical formulae
}
\examples{
SpeedOfSoundSeaWaterSkone(0, 1, 30)
}
\author{
Jose Gama
}
\references{
de Jong, C.D., Lachapelle, G., Skone, S. and Elema, I. A., 2002
Hydrography. Delft University Press (The Netherlands). pp.194
}

