% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_methods.R
\name{[.sonar}
\alias{[.sonar}
\title{Method to subset 'sonar' objects}
\usage{
\method{[}{sonar}(x, i, j, drop = FALSE)
}
\arguments{
\item{i, j}{row and column indices}
}
\value{
'sonar' object
}
\description{
Returns subset of 'sonar' object (data.frame)
}
