% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_Preprocessing.R
\name{summary_func}
\alias{summary_func}
\title{Summary of preprocessing}
\usage{
summary_func(track.ddf.list, track.eval.out.list, merkmale.out.list,
  t.grid.list, n.angle)
}
\arguments{
\item{track.ddf.list}{Output of \code{\link{preprocess.data}}}

\item{track.eval.out.list}{Output of \code{\link{preprocess.data}}}

\item{merkmale.out.list}{Output of \code{\link{preprocess.data}}}

\item{t.grid.list}{Output of \code{\link{preprocess.data}}}

\item{n.angle}{Number of watch hands per quarter, i.e., total number of 
watch hands is \code{4 x n.angle}}
}
\value{
Matrix summarizing results of preprocessing for each object
}
\description{
Summary of preprocessing
}

