% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthCircularPlateNormal}
\alias{TargetStrengthCircularPlateNormal}
\title{target strength Circular Plate normal}
\usage{
TargetStrengthCircularPlateNormal(r, lambda)
}
\arguments{
\item{r}{radius (meters)}

\item{lambda}{wavelength}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Circular Plate normal
}
\examples{
TargetStrengthCircularPlateNormal( 10, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

