% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthConvexSurface}
\alias{TargetStrengthConvexSurface}
\title{target strength Convex surface}
\usage{
TargetStrengthConvexSurface(r1, r2)
}
\arguments{
\item{r1}{principal radii  (meters)}

\item{r2}{principal radii  (meters)}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Convex surface
}
\examples{
TargetStrengthConvexSurface( 100, 50 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

