% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthCylinderNormal}
\alias{TargetStrengthCylinderNormal}
\title{target strength Cylinder normal}
\usage{
TargetStrengthCylinderNormal(r, L, lambda)
}
\arguments{
\item{r}{radius (meters)}

\item{L}{length (meters)}

\item{lambda}{wavelength}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Cylinder normal
}
\examples{
TargetStrengthCylinderNormal( 10, 5, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

