% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthCylinderThetaToNormal}
\alias{TargetStrengthCylinderThetaToNormal}
\title{target strength Cylinder, theta to normal}
\usage{
TargetStrengthCylinderThetaToNormal(r, L, lambda, theta)
}
\arguments{
\item{r}{radius (meters)}

\item{L}{length (meters)}

\item{lambda}{wavelength}

\item{theta}{angle to normal}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Cylinder, theta to normal
}
\examples{
TargetStrengthCylinderThetaToNormal( 10, 5, 500, 45 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

