% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthPlateAnyShape}
\alias{TargetStrengthPlateAnyShape}
\title{target strength Plate of any shape}
\usage{
TargetStrengthPlateAnyShape(A, lambda)
}
\arguments{
\item{A}{area  (meters)}

\item{lambda}{wavelength}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Plate of any shape
}
\examples{
TargetStrengthPlateAnyShape( 10, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

