% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthRectangularPlateNormal}
\alias{TargetStrengthRectangularPlateNormal}
\title{target strength Rectangular Plate normal}
\usage{
TargetStrengthRectangularPlateNormal(A, B, lambda)
}
\arguments{
\item{A}{side, A>=B  (meters)}

\item{B}{side (meters)}

\item{lambda}{wavelength}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Rectangular Plate normal
}
\examples{
TargetStrengthRectangularPlateNormal( 10, 500, 500 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

