% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthRectangularPlateThetaToNormal}
\alias{TargetStrengthRectangularPlateThetaToNormal}
\title{target strength Rectangular Plate, theta to normal}
\usage{
TargetStrengthRectangularPlateThetaToNormal(A, B, lambda, theta)
}
\arguments{
\item{A}{side, A>=B  (meters)}

\item{B}{side (meters)}

\item{lambda}{wavelength}

\item{theta}{angle to normal}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength Rectangular Plate, theta to normal
}
\examples{
TargetStrengthRectangularPlateThetaToNormal( 10, 500, 500, 45 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 70.
}

