% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TargetStrengthSphere}
\alias{TargetStrengthSphere}
\title{target strength sphere}
\usage{
TargetStrengthSphere(r)
}
\arguments{
\item{r}{radius (meters)}
}
\value{
Target Strength (TS) (dB)
}
\description{
Returns the target strength sphere
}
\examples{
PeakTS( 900, 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 69.
}

