% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sonar.R
\name{TransmitDirectivityIndex}
\alias{TransmitDirectivityIndex}
\title{transmit directivity index}
\usage{
TransmitDirectivityIndex(Idir, Iomni)
}
\arguments{
\item{Idir,}{intensity along the axis of the beam pattern}

\item{Iomni,}{intensity of the equivalent non-directional projector}
}
\value{
transmit directivity index (DIt)
}
\description{
Returns the transmit directivity index (DIt)
}
\examples{
TransmitDirectivityIndex( 700, 1000 )
}
\author{
Jose Gama
}
\references{
Waite A. D., 2002
Sonar for Practising Engineers, 3rd Edition
Chichester: Wiley. pp. 4.
}

