% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktionen_alternative_einlesen.R
\name{get.version}
\alias{get.version}
\title{Nur Version bestimmen und win.start / win.length}
\usage{
get.version(file)
}
\description{
Nur Version bestimmen und win.start / win.length
}

