% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonar_funcs.R
\name{sonar_sidescan_geo}
\alias{sonar_sidescan_geo}
\title{Function to georeference sidescan data extracted from sonar}
\usage{
sonar_sidescan_geo(sonar, res = 5e-06, fun = max)
}
\arguments{
\item{res}{Target resolution for grid in degrees.}

\item{fun}{Default = max. Function passed to rasterize.}

\item{'sonar'}{object}
}
\value{
Raster object
}
\description{
Creates georeferenced version of raw sidescan sonar data from XYZ points using raster::rasterize.
Returns output Raster object with sidescan data.
}
